fix(riscv_common): return u32 from csr_read_word for consistent oracle protocol
Revert return type from usize to u32 to ensure the oracle protocol
uses a consistent 32-bit word size on both rv32 and rv64 targets.

+++ riscv_common/src/lib.rs
@@ -14,9 +14,9 @@ pub fn csr_write_word(word: usize) {
 
 #[inline(always)]
 /// Reads a word from CRS register.
-/// Returns usize to preserve full register width on both rv32 and rv64.
-pub fn csr_read_word() -> usize {
-    let mut output: usize;
+/// Returns u32 for consistent oracle protocol on both rv32 and rv64.
+pub fn csr_read_word() -> u32 {
+    let mut output: u32;
     unsafe {
         core::arch::asm!(
             "csrrw {rd}, 0x7c0, x0",
